Move method to separate file
