+++ x-pack/plugins/enterprise_search/public/applications/shared/flash_messages/flash_messages_logic.ts
@@ -80,10 +80,3 @@ export const FlashMessagesLogic = kea<MakeLogicType<IFlashMessagesValues, IFlash
     },
   }),
 });
-
-export const setSuccessMessage = (message: string) => {
-  FlashMessagesLogic.actions.setFlashMessages({
-    type: 'success',
-    message,
-  });
-};
