+++ x-pack/plugins/enterprise_search/public/applications/shared/flash_messages/index.ts
@@ -13,3 +13,4 @@ export {
 } from './flash_messages_logic';
 export { FlashMessagesProvider } from './flash_messages_provider';
 export { flashAPIErrors } from './handle_api_errors';
+export { setSuccessMessage, setErrorMessage, setQueuedSuccessMessage } from './set_message_helpers';
